requirements.txt
alembic==1.14.0
apt-xapian-index==0.49
apturl==0.5.2
asttokens==2.4.1
autoflake==2.3.1
ayatana-settings==21.1.28
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.0.0
caffeine==2.9.8
cairocffi==1.2.0
CairoSVG==2.5.2
cajarename==21.11.24
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
checkbox-support==0.22
click==8.0.3
cmake==3.31.1
colorama==0.4.4
colorlog==6.9.0
comm==0.2.2
command-not-found==0.3
contourpy==1.3.1
cryptography==3.4.8
cssselect2==0.4.1
cupshelpers==1.0
cycler==0.12.1
Cython==0.29.28
dateparser==1.2.0
dbus-python==1.2.18
debugpy==1.8.7
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
deja-dup-caja==0.0.9
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
docopt==0.6.2
duplicity==0.8.21
et_xmlfile==2.0.0
ewmh==0.1.6
exceptiongroup==1.2.2
executing==2.1.0
fasteners==0.14.1
feedparser==6.0.8
filelock==3.6.0
folder-color-caja==0.0.86
folder-color-common==0.0.86
fonttools==4.55.0
fsspec==2025.2.0
func-timeout==4.3.5
fuse-python==1.0.2
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.1.0
gpg==1.16.0
greenlet==1.1.2
guacamole==0.9.2
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
imageio==2.36.1
importlib-metadata==4.6.4
importlib_resources==6.5.2
iotop==0.6
ipykernel==6.29.5
ipython==8.29.0
isort==5.6.4
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.4.2
jsonpath-rw==1.4.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==23.5.0
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
llvmlite==0.43.0
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Magnus==1.0.3
Mako==1.1.3
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mate-hud==22.4.4
mate-tweak==22.4.8
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
meld==3.20.4
menulibre==2.2.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
mysqlclient==1.4.6
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
nibabel==5.3.2
notify2==0.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
onboard==1.4.1
opencv-python==4.11.0.86
openpyxl==3.1.5
optuna==4.1.0
packaging==21.3
padme==1.1.1
pandas==2.2.3
paramiko==2.9.3
parso==0.8.4
patsy==1.0.1
pexpect==4.8.0
pibootctl==0.5.2
pillow==11.0.0
plainbox==0.25
platformdirs==2.5.1
ply==3.11
progressbar2==4.5.0
prompt_toolkit==3.0.48
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pulsemixer==1.5.1
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.44.1
pydicom==3.0.1
PyDrive==1.3.1
pyflakes==3.2.0
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pykwalify==1.8.0
pylibacl==0.6.0
pymacaroons==0.13.0
-e git+https://github.com/bhklab/PymRMRe@87d7a5a150476fbad0846384a843307320e6ac27#egg=pymrmre
PyNaCl==1.5.0
pynndescent==0.5.13
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
pyradiomics==3.0.1
pyRFC3339==1.1
pysmbc==1.0.23
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-docx==1.1.2
python-utils==3.9.1
python-xapp==2.2.1
python-xlib==0.29
pytz==2022.1
pyudev==0.22.0
PyWavelets==1.7.0
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.2.0
radMLBench==1.1
regex==2024.11.6
reportlab==3.6.8
requests==2.25.1
rich==13.9.4
rsa==4.8
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
scikit-image==0.24.0
scikit-learn==1.5.2
scikit-posthocs==0.11.1
scipy==1.14.1
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
setproctitle==1.2.2
sgmllib3k==1.0.0
shellingham==1.5.4
SimpleITK==2.4.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
ssh-import-id==5.11
stack-data==0.6.3
statsmodels==0.14.4
svgwrite==1.4.1
sympy==1.13.1
systemd-python==234
threadpoolctl==3.5.0
tifffile==2024.9.20
tinycss2==1.1.1
toml==0.10.2
tomli==2.1.0
torch==2.6.0
torchvision==0.21.0
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
triton==3.2.0
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ultralytics==8.3.82
ultralytics-thop==2.0.14
umap-learn==0.5.7
unattended-upgrades==0.1
unity-lens-photos==1.0
unity-scope-calculator==0.1
unity-scope-devhelp==0.1
unity-scope-manpages==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-zotero==0.1
unity-tweak-tool==0.0.7
uritemplate==3.0.1
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
xcffib==0.11.1
xdg==5
xdiagnose==3.8.10
xgboost==2.1.2
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==3.2.0
youtube-dl==2021.12.17
zipp==1.0.0
